Split core features and routes into a router so they can be mounted by other apps
